LPD-47626 This is testing the ImportTaskResourceTest
